eliminate header_name type
the only use site was in get_header(), and there the id field wasn't
even used. the latter was arguably a bug, but as header_names and
header_id are in sync anyway, that's fine.

so just convert header_names into a plain string array and codify the
need to keep things in sync.