eliminate smtp_cmd type
the only use sites were in get_id() and smtp_in() (help command), and
there the id field wasn't even used. the latter was arguably a bug, but
as smtp_cmds and smtp_cmd_id are in sync anyway, that's fine.

so just convert smtp_cmds into a plain string array and codify the
need to keep things in sync.