straighten out alias matching
this makes the logic much easier to follow.
as a bonus, we also take the "postmaster" comparison out of the loop
also in the globbing case.

as a drive-by, rename the replacement value to avoid being subsequently
shadowed.

also add a FIXME to the case where domains are incorrectly matched
case-sensitively. fixing this would require parsing the patterns, which
doesn't seem worth it, given that aliases for addresses with a domain
are a corner case (which must be forced with the local_addresses
option).

while we introduce table_find_fnmatch_casefold(), also rename the
previously unused table_find_case() function, as the old name was
ambiguous.