strip whitespace from elements of semicolon-separated lists
subsequent code clearly assumes that there is none, and
parse_list_file() already strips the lines as well.

one could argue that this limits possible values, but a) it already
wouldn't have worked for the first and last element, b) we don't
actually need such values, and c) if we did, some quoting scheme would
be the right approach.

to maintain strtok() semantics, we need to manually skip entries that
become empty due to stripping.

with the pre-stripped list entries, we can do less work in
parse_interface() and parse_table_pair() now.