add support for backslash escapes to address parser
the RFCs call these quoted-pair, but that sounds too unclear to put into
error messages.

also, the RFCs technically require allowing escaping terminating zeros,
but that seems a tad too crazy. we don't expect to see such legacy
mails.

we also fail to dequote the strings for later processing, which may
matter for comparisons. but we don't expect to see such addr-specs,
either.