make parse_address_rfc82[12]()'s address_end parameter optional
for RFC822 it is needed only when parsing address lists.
consequently, we also stop parsing commas when the pointer is NULL, so
we'll complain about unexpected chars.

there is no such thing as a list of RFC821 addresses, so provisions for
list parsing are pointless here. however, we may want to integrate the
address parsing into the SMTP command parsing, at which point we'd need
the out param, and it doesn't cost much to keep it.
for the other case, add a complaint about excess chars.

for _create_address() the parameter already was optional, so the code
becomes simpler due to the logic moving into the lower-level functions.