make parse_address_rfc82[12]()'s address_end parameter optional
for rfc822 it is needed only when parsing address lists.
consequently, we also stop parsing commas when the pointer is NULL.

there is no such thing as a list of rfc821 addresses, so there is no
point in supporting that. however, we may want to integrate the address
parsing into the smtp command parsing, at which point we'd neeed it. it
doesn't cost much to keep it.

for _create_address() it already was optional, so the code becomes
simpler.