fix handling of empty addresses in incoming SMTP streams
the return path is allowed to be empty, and thus we need to exclude that
case from the check that the address is qualified.

the recipient address may not be empty. this was also already rejected
as a side effect of catching unqualified addresses, but it's nicer to
catch it explicitly.