streamline map_address_header() api
construct the header copy inside the function, not at the call site. in
case no rewriting is actually done, that avoids the pointless allocation
and subsequent destruction of the new header object (which also happened
to leak the header payload).

sort of as a drive-by, this fixes creating header objects with an
invalid `value` pointer, though i don't think it is currently being used
subsequently.