don't try to open log/debug files prior to reading config
it makes logically little sense to attempt to do it, because the config
file determines their location. as a consequence, running as non-root
would consistently fail, thus making the integration tests useless.

so just don't try to open any files. error messages will simply go to
the console, or wherever that is redirected to.