convert integration tests to stand-alone shell scripts
the makefiles and templates lead to duplication and hard to follow data
flows. so instead, create the config files in the tests themselves, and
avoid duplication by factoring out a library. this is modeled after the
git autotest suite.

the -oi and -t tests are split from the basic delivery test, because
this improved atomicity now doesn't cost anything.

this drops some still relevant comments from the readmes, which will
become obsolete in subsequent commits. pedantically, this is non-atomic,
but the churn doesn't seem worth it.