ensure at startup that we have sane stdio
stdio is generally assumed to be present, and unrelated fds occupying
that space can lead to "interesting" effects. as a set-uid app, we
better not risk it.

because of that, we can remove the re-opening of the log around closing
stdio in mode_daemon().

note that we cannot just ground stdin & stdout, as some non-daemon
operation modes actually use them.