default to running as mail user instead of root
rather than running as root and changing to the mail user for particular
operations, run as the mail user and acquire root only when necessary.
this is more secure and less code.

note that the -bm and -bs modes now run as the mail user, rather than as
the calling user. this is consistent with -bd, and it shouldn't make a
difference anyway due to all interesting operations already having
switched to the mail user.

is_privileged_user() and set_identity() lose the uid parameter, as it's
always conf.orig_uid. this is not split off to reduce churn.

set_identity() is also renamed to verify_privileged_user() to better
reflect its remaining function (the previous name was sort of misleading
anyway, as the set uid didn't match the passed one).

peidopen() is reduced to peopen() due to the identity-setting function
being unused now.