automate setting package version
this avoids the need for bumping the version, which is particularly
annoying if one doesn't know yet whether the next release will be a
patch, minor, or major.

the authoritative source is git, but we also have a fallback for the
case of somebody rebuilding configure from an extracted tar-ball.

note that it's impossible to determine the version at configure
time, so you need to remember to run autoconf prior to rolling a
tar-ball.