replace peopen() with glib functions
use g_spawn_async*() to launch the processes, and g_shell_parse_argv()
to parse the command lines.

as a side effect, quoting is now supported in the commands. use this to
make the integration tests (more) resilient to paths with spaces.

another side effect is that the number of command line arguments is not
limited to 10 anymore.

either of the two above causes the test mail delivered by t132-local-mda
to actually have the entire intended content, but this isn't checked.

while we update the docu, also make explicit that env vars are not
supported in the arguments, rather than implying that they are, as was
the case before v0.1.8.