improve mx record resolution
per rfc 5321 section 5.1, it is explicitly forbidden for mx records to
point to cnames, so remove support for that. n.b., the opposite case of
cnames pointing to mx records is permitted, so we may want to add
support for that later.

secondly, the hosts obtained from mx records may be multi-homed, i.e.,
they may result in multiple 'a' records, so handle that case.