absorb resolve_list dns_a option into dns_mx
the smtp spec proscribes that mx records should be used when found, with
a straight fallback to 'a' records. there is no reason to leave setting
up that fallback chain to the user. and using just 'a' records isn't of
much use, as the 'byname' option would cover such cases (if they exist).