improve error propagation from smtp_out_open*()
create the error messages closer to the error sources, rather than in
deliver_msglist_host_smtp().

as a drive-by, we make the reporting of wrapper errors less talkative -
the sender of a message really shouln't know what exactly is wrong with
the MTA config, unless they are the admin and thus have access to the
logs anyway.