optimize alias_expand() api
instead of passing back the resolved address list and letting the caller
replace the original, do the replacement within the function, making the
code less noisy.

we also do that for the recursively called expand_one() function.

while we're changing function signatures, also reshuffle the argument
order to be more logical.