fix handling of unqualified addresses
the code didn't agree with itself whether "unqualified" is represented
by NULL or an empty string: NULL was never created, but all conditionals
tested for it and thus were broken.

the choice which one to pick is somewhat arbitrary, but avoiding NULL
yields safer code while being less verbose.

the "paranoid check" in spool_write_rcpt() is moved to create_address().
it is really an assert, and there is no reason to try to recover from
it.