fix & simplify route recipient filtering
don't split the list into two lists and then leak one of them.

in that process, route_split_rcpts() gets re-purposed as
route_filter_rcpts(). while we're at it, we also move it up a bit, so
it's adjacent to its worker filter_rcpts(), and
route_{sender,from_hdr}_is_allowed() are finally adjacent as well.

as a drive-by, the search loop in filter_rcpts() now uses a more
conventional pattern.

we also move the call in deliver_route_msg_list() a bit up, because this
better fits the structure: first filter the list, then do the boolean
checks. the 'From:' check is most expensive, so it should be last.