fix memory leaks of recipient objects
it turned out to be no fun at all to track the ownership of the objects:
msgout gets a shallow copy of its msg's list, but then mixes in newly
allocated objects from alias resolution (and the log_user for good
measure). pointers to objects from this list are then selectively
appended to the msg's non_rcpt_list, which may already contain objects
allocated when reading the spool. finally, the parent-child relations of
expanded aliases need to be owning, because the children do not end up
in an rcpt_list.

so after much deliberation i took the easy way out and introduced
ref-counting.