clean up line wrapping/indentation somewhat
this optimizes the layout for max 100 columns with a tab width of 4.

tabs are used for indentation, and spaces for padding/lineup. however,
to limit overhangs, i often use two indents instead of padding - the
decisions are made based on visual context.

to limit the churn somewhat, this omits cases where the only change
would be replacing indentation with padding, without significantly
improving legibility.