re-purpose address.address field
so far, the field held the complete address including display name,
angle brackets, and comments.

but all (remaining) uses of the field are in debug messages and SMTP 250
OK messages, which don't actually need the extra gunk.

therefore, re-purpose the field as a cache for the actual addr-spec,
which will allow some optimization and cleanup. the creation of the
address object itself is also simplified.