stop abusing errno for communicating sysexits
after the error message propagation cleanup, we can do that.

due to the removed "multiplexing", ext_strerror() can be simplified to
sysexit_str().

as a drive-by, add some static asserts around the sysexit message string
array to ensure that we're looking up the right thing, and make use of
G_N_ELEMENTS().