match globbing and regular aliases in one go
we want to retain easy access to the original address in case expansion
fails, which we'll need later. multi-stage lookup is in the way of that.

note that this has the side effect of making evaluation interleaved
instead of sequential: a regular alias may now be expanded by a glob in
turn.