dynamically determine path for recursive invocation
otherwise it doesn't work with non-installed builds, which is bad for
testing.

the implementation uses the whereami library, which was the most
generic answer found on stackoverflow. it does more than we need and
it's a bit of a mess, but i'm not going to touch code with so many
platform ifdefs.