don't enforce existence of local users when run_as_user
as we cannot set the ownership of the mailbox anyway (unless we run as
root, but that's not the intended use case), it doesn't matter whether
it belongs to a real user or not.

this allows arbitrarily named mailboxes, which we take advantage of in
the integration tests by using distinct fixed strings instead of the
local user for the addresses' local parts.