make deliver_msg_list() call chain void
as indicated by the comments, there was confusion about the meaning of
the return value. sometimes it meant that all messages were delivered,
other times that some were delivered, and in yet other cases it was just
garbage.

in the mode_accept()/smtp_in() case it was kind of ignored anyway,
because our sigchild handler would just log the event (which was
redundant with the more specific messages logged near the problem
source), and not do anything else - which is just fine, because we'll
retry later anyway.

in the run_queue() case the exit code actually reflected the value, but
it was undocumented, and one couldn't really deduce anything useful from
it anyway.