rework pruning of already delivered alias-expanded addresses
it's not necessary to post-process the recipient list after alias
expansion, as we subsequently split the list anyway, where we can just
skip non-recipients.

it would be faster to not put them into the tree in the first place, but
that would require integrating the filtering with the alias expansion,
which would make the code messier for negligible gain.