overhaul GList iteration loops
make the foreach macro a lot more powerful, and use it throughout.

its arguments are swapped for familiarity - with Qt, C++ range-for(),
and even the plain for().

also add a mutating variant.
note that some use sites don't actually modify the node links, but only
the payloads, so using this macro is less efficient than using one with
only a non-const iterator. but adding that would be really just clutter
for marginal gain.

this makes a few closely related drive-by changes:
- the search loop in route_msgout_list() now uses a more conventional
  pattern
- a lot of unnecessary uses of g_list_first() are dropped, most via
  foreach itself
- in deliver_msglist_host_smtp(), we use ->next instead of g_list_next()