expand variables only after splitting pipe/MDA commands
that way, the user doesn't need to quote the variable expansions, and
malicious values cannot mess with the command line parser. the latter
was a potential security issue, as a remote sender could have caused
additional arguments to be passed to the command - probably not as bad
as outright shell injection as fixed in v0.2.21, but still.