don't fork() for delivery by default
making delivery async merely makes the smarthost potentially refuse to
cooperate, as we might try to open too many connections as the same
time. this would happen for example when git-send-email is used on a
somewhat longer patch series.

additionally, returning before the delivery attempt completes makes
testing more complex and slower, as we need to keep retrying until a
timeout.

an alternative approach to fully sync operation would be managing
outgoing connections, but that would be major overkill for us, and
wouldn't help the testing case anyway.

the previously ignored -odb option now actually has an effect, to force
the old behavior.