bequeath our stderr to helper processes
the connection wrapper and pipe commands (incl. pipe aliases and MDA
commands) now inherit stderr, so it ends up in our log file. arguably,
it would be better to relay the output to syslog when we're using that,
but this would require async process handling.

there is no security risk associated with bequeathing stdio, as the
other ends don't belong to us, so we can't leak anything to the child.
there is a risk that the child would fill up the disk, but our own
logging is not internally rate-limited, either.

implementation-wise, for the connection wrapper command we simply stop
dup2()'ing the pipe to stderr in child(). for the other helpers, we
don't close most of stdio in peopen() anymore - that was just wrong in
the first place, as starting processes without valid stdio is bad, even
if the particular descriptors don't interest us.

we remove the /dev/null redirection from the openssl wrapper examples,
as it is redundant now. but we add -verify_quiet, as -quiet doesn't
suppress the certificate-related messages.